rename DatetimeRow 2 GeneratedDatetime & relocate
