--- domains/rem/src/ui/generatedDates/DatetimeRow/styles.scss
@@ -1,15 +1,0 @@
-.ee-datetime-row {
-	&__title {
-		align-items: center;
-		display: inline-block;
-		padding: var(--ee-padding-tiny);
-
-		svg {
-			margin-right: var(--ee-margin-tiny) !important;
-		}
-
-		span {
-			color: white;
-		}
-	}
-}
